working [read long]
works normally with disconnections, connections & unexpected interruptions (in which it may exit after a timeout of 15s on udp_recv). the number of POPs influences in someway the gathering of POPs but in reality, each new connection means a new PQ in order to ENSURE POP FRESHNESS. Any ideas to solve this problem without exhaustive communication using only the protocols with the headers described in recv_upstream and recv_downstream are welcome.